refactor: move ServiceContainer from domain/ to application/ layer
Fix hexagonal architecture violation where ServiceContainer in domain/
imported from infrastructure/ and application/. The domain layer should
never depend on outer layers.

- Keep only IServiceContainer, ServiceKey, ServiceFactory, and T TypeVar
  in domain/container/service_container.py (pure interface)
- Move ServiceContainer, ImmutableServiceContainerBuilder, and factory
  functions to application/container/service_container.py (concrete impl)
- Update all imports across app_factory, routes, application context,
  domain factories, and test files
