+++ application/container/__init__.py
@@ -1,0 +1,1 @@
+# Application-layer service container implementation
